+++ backend/src/main/java/com/prem/ta/models/OrderFlowMetrics.java
@@ -1,7 +1,9 @@
 package com.prem.ta.models;
 
+import java.math.BigDecimal;
+
 public record OrderFlowMetrics(
-        double buyerVolumeShare,
-        double buyerCapitalShare
+        BigDecimal buyerVolumeShare,
+        BigDecimal buyerCapitalShare
 ) {
 }
